sdk/client: top-level byte streams should always be pointers to pointers
Within the SDK we have a convention to output a byte stream as a pointer within a
wrapper model struct e.g.

type ExampleOperationResult struct {
  Model *[]byte
}

which is intentionally always a pointer since no response body maybe present.

As such we can assume that the object we unmarshal into is always a **byte