sdk/client/resourcemanager: parsing error messages from API Responses
This commit solves the first part of a number of issues within `hashicorp/terraform-provider-azurerm`
by providing a function which parses the Azure Error from the API Response
for both Long Running Operations and regular Resource Manager operations.

There's a considerable number of different formats for error messages, so this isn't
going to be necessarily be conclusive, hence falling back to a generic error message
showing the API body if we can't parse it.