auth: ability to manually expire a cached token
Whilst the new ExpireToken() method only has any meaning or effect when
a token is being cached, in practise this all the existing Authorizers wrap
themselves with a CachedAuthorizer. For consistency this is added to the
Authorizer interface to make downstream implementation simpler and this
should not cause any incompatibility or errors where one of the provided
constructors is used.